Caesar Cipher - Cryptography
In order to solve the problem of creating an encryption and decryption message, I utilized the chr () and ord () functions. This method was used to leverage the Unicode standard which represents a list of every character used by the human language and assigns each character its unique code. This method is effective because Unicode can be used globally which allows computers to recognize it regardless of if the text entered is in another language besides English. The ord () function is used to convert any character to a numeric value that python (the computer) can interpret and the chr() function does the exact inverse, finding the number and converting it into the corresponding character. This is the best method to switch between encrypted and decrypted messages. I first defined the parameters. Throughout my code I used the following variables to represented and described below:
•	User_txt = the text the user submitted
•	wip_index = internal reference that returns the position of listed elements
•	wip_scramble = function that ultimately creates the ciphered message by combining the index variable and the key provided by the user
•	wip_newscramble = the variable that represents a new character
•	wip_org = the original position of the listed element
•	wip_org_pos= right/positive shift that allows the key to move to the next letter or Unicode

We first defined & separated the functions of encryption and decryption using def to store our function to be used later in our print commands. We want to repeat the code and define the number of repetitions to continuously covert the unicodes, so the for loop is used. In addition to the for loop we will also use conditional statements to create the parameters. The conditional statement used is if, elif, and else. These commands essentially directs the text inputs and outputs to follow a specific rule based upon the conditions set. As previously stated, the ord function is used to convert any character to numeric value and the conditions were based upon if the character were lower case or upper case. We created the index for the ord to reference and added it with the key variable that the user will later provide. Then we created the scrambled text in the encryption function by using the % operator which divides and returns the value of the remainder. We used 26 as an integer to be divided because we are using all 26 letters of the English alphabet. We then multiplied or stored the wip_scrambled variable with the chr function to create the new scrambled variable. We then used the += operator to add the encrypted variable with the newly created scrambled variable. This process was repeated for upper case letters, the only difference is that it is for uppercase. For the decryption function we mirrored the encryption function with setting the conditional statements for lower and uppercase;however, the decryption function introduces new variables. The new variables signify the original position of the Unicode characters. The return function was used to close both of the encryption and decryption loops. 

Finally, we created lines for user inputs. Our key, which was referenced and stored in the encryption and decryption functions, requites an input from the user. Therefor, the code can be used to not only shift the Unicode position 3 letters, but I can also move towards any entered integer because of the int function. The message variable also requires user input. Lastly, we used the print function and referenced the defined paraments of the getEncryption and get Decryption functions.