RSA - Cryptography
PROMPT

Implement RSA by creating specific core functions: (a) key generation and (b) encryption/decryption, should
be implemented. 

Pre/post conditions of the functions are as follows:
- KeyGen (p, q) -> {e, n}, {d, n}: This function takes as input two primes then outputs
public/private keys. There is no restriction for the type of input/output.
- RSA (k, M) -> This function takes as input a key (private or public) and a message (or
a ciphertext) then outputs a ciphertext (or a plaintext)
For the simplicity, the following assumptions should be applied:
• The KeyGen function takes two primes, both of them are smaller than 100 (no need to
check whether they are primes).
• The KeyGen function should randomly select the private key ‘e’ in the range of 3 and
𝜙𝜙(𝑛𝑛).
• A brute-force approach can be used to find a multiplicative inverse (no need to
implement extended Euclidean Algorithm).
• The RSA function does not have to check the type of input, which means we do not
care the input is a ciphertext or a plaintext.
• In case of encryption, RSA function takes as input an integer range from 0 to 256 then
outputs a ciphertext in an integer form. It can be extended to take a list of integers,
which is optional. 


SOLUTION explained:In order to solve the problem of creating an RSA encryption/decryption message, I utilized the import python import libraries. This method used both the math and random libraries for math operations for gcd. GCD stands for greatest common denominator, and it is applied to the r (n) value which is then added to the coprime array. The coprime array consists of paired uniform linear array which provides coprime numbers. This is the best method to switch between encrypted and decrypted messages. I first defined the parameters. Throughout my code I identifies the following variables that were most important to the construction of the RSA message, represented and described below:
•	r : apart of the key gen function
•	coprimes : array for prime integers
•	k_gen : uses the imported math library to find the GCD of r and the values are added to the coprime array
•	ran : random.choice :random library used to pick on random value from the coprime array
•	manual: user input variable
We first began the implementation with importing the math and random library which provides us with the math functions and the random library selects the correct variable for r. Due to the structure of my code, I decided it would be best to separate the entry of the two prime numbers instead of using the .split() function.  We moved forward to defining our libraries and using the for range function to satisfy the requirement for the private key to stay within the range of 3. We then defined and coprimes and random library and created essentially the decryption function that also references they math library using phIN. Lastly, I created the print function to display the public key for encryption which is the first number provided by the public key and private key. We entered the numbers for the cipher and produced the plain text