Average Grades
PROMPT: Create a solution in python that allows the user to input 8 numerical grades and the system converts their entries into a letter grade

SOLUTION explained: In order for me to convert an integer to a letter grade I need to create a program that requests the customer to input grades 8 times. I defined the calc_average function with the following variables: 
-num: numbers that will be input by the user
-grade: the sum of num and the score
-avg: grade divided by 8 to take the average of all the inputs

I created a sequence/loopback of the input from 1 to 9 which will allow the input function to iterate 8 times. I also defined both the num and grade variable by formatting the input value as an integer and the grade variable was defined by adding the num. The avg grade is finally divided by 8 since we are determining the mean value. I defined the next function to determine the grade score and I used an if...else statements since the condition if the value of each letter grade was already specified. 
Lastly, I broke the previous function called with return and created my last print function to specify the final grade.