Cat Sim
PROMPT: Create a program where the end-user interacts with a cat. In this scenario, the end-user should be able to enter the name of the cat, interact with the cat with two different options, "play" or "feed".

SOLUTION: I created two different py projects where the main file imported another python module, cat. On the main module, I defined the main module, created the input function for the user to determine the cats name, specified the cat's name based upon the users' input, and specified the variables for the conditional if/else loop back for the Y and N values. I created the while true conditional for yes and created a task for both options. 

On a different python document entitled, cat, I defined the initial interaction and specified variables for each interaction and stored the values.